+++ KS.Entities/Cart.cs
@@ -9,18 +9,15 @@ namespace KS.Entities
         public string CustomerId { get; set; }
         public ApplicationUser Customer { get; set; }
 
-        public int ProductId { get; set; }
+        public long ProductId { get; set; }
         public Product Product { get; set; }
 
         public int Quantity { get; set; }
         public decimal Price { get; set; }
+
         public decimal TotalPrice
         {
-            get
-            {
-                return Quantity * Price;
-            }
+            get { return Quantity * Price; }
         }
-
     }
 }
