+++ KS.Entities/StockHistory.cs
@@ -2,11 +2,11 @@
 {
     public class StockHistory : BaseEntity
     {
-        public int ProductId { get; set; }
+        public long ProductId { get; set; }
 
         public Product Product { get; set; }
 
-        public int WarehouseId { get; set; }
+        public long WarehouseId { get; set; }
 
         public Warehouse Warehouse { get; set; }
